refactored startup sequence to auto-detect runtime environment. plus help usage.
