+++ src/main/java/com/logdyn/ui/console/commands/repository/RepositoryCommand.java
@@ -5,9 +5,4 @@ import picocli.CommandLine.Command;
 import picocli.CommandLine.HelpCommand;
 
 @Command(name ="repository", description = "List, add, or rename repositories", subcommands = HelpCommand.class)
-public class RepositoryCommand extends CliCommand {
-    @Override
-    public void run() {
-
-    }
-}
+public class RepositoryCommand extends CliCommand {}
