+++ src/main/java/com/logdyn/ui/javafx/FxApplication.java
@@ -1,6 +1,8 @@
 package com.logdyn.ui.javafx;
 
 import javafx.application.Application;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.stage.Stage;
 import org.apache.log4j.Logger;
 
@@ -11,6 +13,7 @@ public class FxApplication extends Application {
     @Override
     public void start(final Stage primaryStage) throws Exception {
         LOGGER.debug("Starting JavaFX Application...");
+        new Alert(AlertType.INFORMATION, "blarg").showAndWait();
         this.stop();
     }
 }
